+++ apps/active_editor.js
@@ -1,6 +1,7 @@
 /**
  * @Author Nikola Lukic
  * @Description Matrix Engine Api Example.
+ * How to use texture editor
  */
 
 import App from "../program/manifest";
@@ -18,7 +19,10 @@ export var runThis = world => {
 
   // eslint-disable-next-line no-unused-vars
   var tex = {
-    source: ["res/images/complex_texture_1/diffuse.png"],
+    source: [
+      "res/images/complex_texture_1/diffuse.png",
+      "res/images/semi_pack/gradiend_half2.png"
+      ],
     mix_operation: "multiply",
   };
 
@@ -41,19 +45,16 @@ export var runThis = world => {
     }
   });
 
+  // Matrix Engine use visual-js game engine
   world.Add("cubeLightTex", 1, "outsideBox", tex);
-
   App.scene.outsideBox.rotation.rotz = -90
   App.scene.outsideBox.position.y = 0;
   App.scene.outsideBox.position.z = -55;
-  // App.scene.outsideBox.rotation.rotationSpeed.z = 50;
-  // App.scene.outsideBox.rotValue = 90;
   App.scene.outsideBox.LightsData.ambientLight.set(1, 1, 1);
   App.scene.outsideBox.glBlend.blendEnabled = true;
   App.scene.outsideBox.glBlend.blendParamSrc = ENUMERATORS.glBlend.param[6];
   App.scene.outsideBox.glBlend.blendParamDest = ENUMERATORS.glBlend.param[7];
-  App.scene.outsideBox.rotation.SetDirection(1, 1, 0.5);
-
+  // App.scene.outsideBox.rotation.SetDirection(1, 1, 0.5);
   // CANVAS2D_SURFACE - IS TEXTURE EDITOR
   E("HOLDER_STREAMS").style.display = "block";
   E("webcam_beta").style.display = "none";
@@ -61,7 +62,6 @@ export var runThis = world => {
     "http://localhost/PRIVATE_SERVER/me/me/2DTextureEditor/actual.html",
     "actualTexture"
   );
-
 
   // Walls
   world.Add("cubeLightTex", 1, "WALLRIGHT", texStone);
@@ -71,7 +71,7 @@ export var runThis = world => {
   App.scene.WALLRIGHT.position.SetX(14);
   App.scene.WALLRIGHT.position.SetY(3);
   App.scene.WALLRIGHT.position.SetZ(-15);
-
+  App.scene.WALLRIGHT.geometry.setTexCoordScaleXFactor(1);
 
   world.Add("cubeLightTex", 1, "WALLLEFT", texStone);
   App.scene.WALLLEFT.geometry.setScaleByX(0.5);
@@ -80,6 +80,7 @@ export var runThis = world => {
   App.scene.WALLLEFT.position.SetX(-14);
   App.scene.WALLLEFT.position.SetY(3);
   App.scene.WALLLEFT.position.SetZ(-15);
+  App.scene.WALLLEFT.geometry.setTexCoordScaleYFactor(1);
 
   world.Add("cubeLightTex", 1, "WALLFRONT", texStone);
   App.scene.WALLFRONT.geometry.setScaleByX(14);
@@ -88,14 +89,13 @@ export var runThis = world => {
   App.scene.WALLFRONT.position.SetX(0);
   App.scene.WALLFRONT.position.SetY(3);
   App.scene.WALLFRONT.position.SetZ(-27);
-  // App.scene.WALLFRONT.rotation.rotx = 90;
+  App.scene.WALLFRONT.geometry.setTexCoordScaleXFactor(1);
 
-  // Load Physics world!
+  // Load Physics world
   let gravityVector = [0, 0, -9.82];
   let physics = world.loadPhysics(gravityVector);
-  // Add ground
+  // Add ground (this func create scene object FLOOR_STATIC)
   physics.addGround(App, world, tex);
-
   var b = new CANNON.Body({
     mass: 5,
     position: new CANNON.Vec3(0, -5, 1),
@@ -105,8 +105,7 @@ export var runThis = world => {
   // Physics
   App.scene.outsideBox.physics.currentBody = b;
   App.scene.outsideBox.physics.enabled = true;
-
-  // setTimeout(function () {
   App.scene.outsideBox.streamTextures.showTextureEditor();
-  // }, 100);
+
+  App.scene.FLOOR_STATIC.geometry.setTexCoordScaleFactor(1);
 };